fix：push API沒有寫入資料庫  問題： 1.  push API沒有寫入資料庫 原因： 1.  api路徑沒有寫好 調整項目：         modified:   public/js/services/notifications.js

+++ public/js/viewHandlers.js
@@ -273,6 +273,11 @@ async function toggleNotification(currentNotificationMethod) {
 function loadNotifications(currentNotificationMethod) {
   const token = localStorage.getItem("token");
 
+  if (!token) {
+    window.location.href = "/";
+    return;
+  }
+
   fetch(`/api/track?notificationMethod=${currentNotificationMethod}`, {
     // 使用 GET 方法和查詢參數
     method: "GET",
@@ -281,7 +286,16 @@ function loadNotifications(currentNotificationMethod) {
       Authorization: token,
     },
   })
-    .then((response) => response.json())
+    .then((response) => {
+      if (!response.ok) {
+        if (response.status === 401) {
+          window.location.href = "/"; // 如果響應是 401，重定向到登錄頁面
+          return;
+        }
+        throw new Error("無法獲取通知");
+      }
+      return response.json();
+    })
     .then((data) => {
       const notificationHeader = document.getElementById(
         `notificationHeader-${currentNotificationMethod}`
@@ -340,6 +354,11 @@ function deleteNotification(notificationId, currentNotificationMethod) {
   alert("確定要刪除嗎？");
   const token = localStorage.getItem("token");
 
+  if (!token) {
+    window.location.href = "/";
+    return;
+  }
+
   fetch(`/api/track?id=${notificationId}`, {
     // 使用 DELETE 方法和查詢參數
     method: "DELETE",
@@ -349,7 +368,17 @@ function deleteNotification(notificationId, currentNotificationMethod) {
     },
   })
     .then((response) => {
-      if (response.ok) {
+      if (!response.ok) {
+        if (response.status === 401) {
+          window.location.href = "/"; // 如果響應是 401，重定向到登錄頁面
+          return;
+        }
+        throw new Error("無法獲取通知");
+      }
+      return response.json();
+    })
+    .then((data) => {
+      if (data.ok) {
         loadNotifications(currentNotificationMethod); // 重新加載通知，更新列表
       } else {
         throw new Error("刪除失敗");
@@ -364,13 +393,18 @@ function toggleFavorite(symbol, favoriteButton) {
     alert("登入後即可使用此功能");
     return;
   }
+  const token = localStorage.getItem("token");
+
+  if (!token) {
+    window.location.href = "/";
+    return;
+  }
 
   // 檢查狀態
   const isFavorite = favoriteButton.classList.contains("btn-warning");
 
   const apiEndpoint = "/api/favorite";
   const method = isFavorite ? "DELETE" : "POST";
-  const token = localStorage.getItem("token");
 
   const queryParams = new URLSearchParams({ userId, symbol }).toString();
   const url = isFavorite ? `${apiEndpoint}?${queryParams}` : apiEndpoint;
@@ -385,6 +419,11 @@ function toggleFavorite(symbol, favoriteButton) {
   })
     .then((response) => {
       if (!response.ok) {
+        if (response.status === 401) {
+          // 401 表示 Unauthorized
+          window.location.href = "/"; // 重新導向到登錄頁面
+          return;
+        }
         throw new Error("無法更新追蹤清單");
       }
       // 切换按鈕狀態
@@ -404,13 +443,29 @@ function toggleFavorite(symbol, favoriteButton) {
 function loadFavorites() {
   return new Promise((resolve, reject) => {
     const token = localStorage.getItem("token");
+
+    if (!token) {
+      window.location.href = "/";
+      return;
+    }
+
     fetch("/api/favorite", {
       method: "GET",
       headers: {
         Authorization: token,
       },
     })
-      .then((response) => response.json())
+      .then((response) => {
+        if (!response.ok) {
+          if (response.status === 401) {
+            // 401 表示 Unauthorized
+            window.location.href = "/"; // 重新導向到登錄頁面
+            return;
+          }
+          throw new Error("無法獲取追蹤清單");
+        }
+        return response.json();
+      })
       .then((data) => resolve(data.favorites || []))
       .catch((error) => reject(error));
   });
@@ -488,6 +543,12 @@ function updateFavoritesModal() {
 
 function removeFavorite(symbol, userId) {
   const token = localStorage.getItem("token");
+
+  if (!token) {
+    window.location.href = "/";
+    return;
+  }
+
   const queryParams = new URLSearchParams({ userId, symbol }).toString();
   const url = `/api/favorite?${queryParams}`;
 
@@ -500,8 +561,16 @@ function removeFavorite(symbol, userId) {
   })
     .then((response) => {
       if (!response.ok) {
-        throw new Error("Failed to remove favorite");
+        if (response.status === 401) {
+          // 401 表示 Unauthorized
+          window.location.href = "/"; // 重新導向到登錄頁面
+          return;
+        }
+        throw new Error("無法獲取追蹤清單");
       }
+      return response.json();
+    })
+    .then((data) => {
       updateFavoritesModal(); // 重新載入列表
 
       // 更新追蹤按鈕狀態
